fall_detection/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow
mediapipe
python-multipart
